requirements.txt
asgiref==3.4.1
Django==3.2.8
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
